+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### Features / Enhancements
 
-- Update plugins included in Grafana Marketplace (#18)
+- Update plugins included in Grafana Catalog (#18)
 - Update to Grafana 9.1.0 (#19)
 - Update to Grafana 9.1.1 (#20)
 - Add Balena Application YouTube video (#20)